refactor: 네트워크 서비스의 프로토콜 / 기본 클래스 이름 변경 및 파일 분리
